Draft: move more functions to draftgeoutils.faces
